Migrate victory-legend to TypeScript
