+++ packages/victory-legend/package.json
@@ -27,6 +27,9 @@
   "peerDependencies": {
     "react": ">=16.6.0"
   },
+  "devDependencies": {
+    "victory-legend": "*"
+  },
   "publishConfig": {
     "provenance": true
   },
