--- packages/victory-legend/src/index.js
@@ -1,1 +1,0 @@
-export { default as VictoryLegend } from "./victory-legend";
